fix override functions
