+++ wrapper/godot-library-extension/src/main/kotlin/godot/core/helpers/helpers.kt
@@ -1,0 +1,2 @@
+package godot.core.helpers
+
